model/requirements.txt
Flask
flask_restful
Pillow